+++ src/main/java/cz/cvut/fit/ejk/gaidumax/drive/controller/FolderController.java
@@ -3,6 +3,8 @@ package cz.cvut.fit.ejk.gaidumax.drive.controller;
 import cz.cvut.fit.ejk.gaidumax.drive.dto.FolderDto;
 import cz.cvut.fit.ejk.gaidumax.drive.mapper.FolderMapper;
 import cz.cvut.fit.ejk.gaidumax.drive.service.interfaces.FolderService;
+import cz.cvut.fit.ejk.gaidumax.drive.service.security.interfaces.AuthService;
+import io.quarkus.security.Authenticated;
 import jakarta.inject.Inject;
 import jakarta.validation.Valid;
 import jakarta.ws.rs.Consumes;
@@ -25,12 +27,15 @@ public class FolderController {
     FolderService folderService;
     @Inject
     FolderMapper folderMapper;
+    @Inject
+    AuthService authService;
 
     @Deprecated(forRemoval = true)
     @GET
     @Path("/{id}")
     @Produces(MediaType.APPLICATION_JSON)
     public FolderDto get(@PathParam("id") UUID id) {
+        authService.checkUserHasAccessToFolder(id);
         var folder = folderService.getByIdOrThrow(id);
         return folderMapper.toDto(folder);
     }
@@ -38,7 +43,9 @@ public class FolderController {
     @GET
     @Path("/{id}/parents")
     @Produces(MediaType.APPLICATION_JSON)
+    @Authenticated
     public List<FolderDto> getAllParentFolder(@PathParam("id") UUID id) {
+        authService.checkUserHasAccessToFolder(id);
         var folders = folderService.getAllParentFolders(id);
         return folderMapper.toDtos(folders);
     }
@@ -46,6 +53,7 @@ public class FolderController {
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
+    @Authenticated
     public FolderDto create(@Valid FolderDto folderDto) {
         var folder = folderService.create(folderDto);
         return folderMapper.toDto(folder);
@@ -55,14 +63,18 @@ public class FolderController {
     @Path("/{id}")
     @Consumes(MediaType.APPLICATION_JSON)
     @Produces(MediaType.APPLICATION_JSON)
+    @Authenticated
     public FolderDto update(@PathParam("id") UUID id, @Valid FolderDto folderDto) {
+        authService.checkUserHasAccessToFolder(id);
         var folder = folderService.update(id, folderDto);
         return folderMapper.toDto(folder);
     }
 
     @DELETE
     @Path("/{id}")
+    @Authenticated
     public void update(@PathParam("id") UUID id) {
+        authService.checkUserHasAccessToFolder(id);
         folderService.delete(id);
     }
 }
